+++ CHANGES.rst
@@ -10,6 +10,8 @@ Version 2.2, 2015-??-??
 - Improved default README.rst, issue #51
 - Use tests/conftest.py instead of tests/__init__.py, issue #52
 - Use setuptools_scm for versioning, issue #43
+- Require setuptools>=9.0, issue #56
+- Do not create skeleton.py during an update, issue #55
 
 .. note::